+++ deps.ts
@@ -36,9 +36,8 @@ export type { WalkEntry } from "https://deno.land/std@0.109.0/fs/walk.ts";
 export { Marked as md } from "https://deno.land/x/markdown@v2.0.0/mod.ts";
 export { parse as parseFlags } from "https://deno.land/std@0.110.0/flags/mod.ts";
 
-import * as log from "https://deno.land/std@0.109.0/log/mod.ts";
-export { log };
-export type Logger = typeof log;
+export { getLogger } from "https://deno.land/std@0.137.0/log/mod.ts";
+export type { Logger, handlers } from "https://deno.land/std@0.137.0/log/mod.ts";
 
 export { serve } from "https://deno.land/std@0.121.0/http/server.ts";
 export { debounce } from "https://deno.land/std/async/mod.ts";